requirements.txt
fastapi==0.111.0
pydantic-settings==2.2.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.7.1
alembic==1.13.1
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.6
python-multipart==0.0.9
